Refactor: streamline Obras management by removing redundant components and enhancing search functionality
