+++ src/components/shared/ObraSearchWrapper.tsx
@@ -2,14 +2,7 @@
 
 import { useState, useEffect } from 'react'
 
-function useDebounce<T>(value: T, delay: number): T {
-  const [debounced, setDebounced] = useState(value)
-  useEffect(() => {
-    const handler = setTimeout(() => setDebounced(value), delay)
-    return () => clearTimeout(handler)
-  }, [value, delay])
-  return debounced
-}
+import useDebounce from '@/hooks/useDebounce'
 
 export default function ObraSearchWrapper({
   onSearch,
